fix: break onboarding segment dependency cycle
